requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
langchain-google-genai==2.0.1
langchain==0.3.3
langchain-chroma==0.1.4
protobuf==3.20.3
langchain-community==0.3.2
python-dotenv==1.0.1
pypdf==5.0.1
google-ai-generativelanguage==0.6.10
google-generativeai==0.8.3
PyPDF2==3.0.1
Flask==3.0.3
langchain-experimental==0.3.2
gunicorn
langchainhub==0.1.21
sentence-transformers==3.2.0
